% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_fragility.R
\name{continuous_fragility_index_vec}
\alias{continuous_fragility_index_vec}
\title{Vectorised Continuous Fragility Index}
\usage{
continuous_fragility_index_vec(
  mean1,
  sd1,
  n1,
  mean2,
  sd2,
  n2,
  conf.level = 0.95,
  n_sim = 5L,
  tol_mean = 0.01,
  tol_sd = 0.01
)
}
\arguments{
\item{mean1, sd1, n1, mean2, sd2, n2}{Numeric vectors of summary statistics.}

\item{conf.level, n_sim, tol_mean, tol_sd}{See \code{\link[=continuous_fragility_index_summary]{continuous_fragility_index_summary()}}.}
}
\value{
A numeric vector of CFI values.
}
\description{
Vectorised wrapper around \code{\link[=continuous_fragility_index_summary]{continuous_fragility_index_summary()}} for use
inside \code{dplyr::mutate()}.
}
